# app-backend/docs/swagger.yaml
basePath: /
definitions:
...
  p2p.createOfferRequest:
    properties:
      asset:
        type: string
      availableLiquidity:
        type: string
      country:
        type: string
      countryCode:
        type: string
      currency:
        type: string
      maxOrderAmount:
        type: string
      merchantPayoutAddress:
        type: string
      minOrderAmount:
        type: string
      offerType:
        type: string
      paymentMethodId:
        type: string
      price:
        type: string
      priceMargin:
        type: string
      priceType:
        type: string
      remark:
        type: string
    type: object
...
  p2p.updateOfferRequest:
    properties:
      availableLiquidity:
        type: string
      maxOrderAmount:
        type: string
      merchantPayoutAddress:
        type: string
      minOrderAmount:
        type: string
      paymentMethodId:
        type: string
      price:
        type: string
      priceMargin:
        type: string
      priceType:
        type: string
      remark:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: Core wallet API for the Trovo platform — user accounts, crypto assets,
    deposits/withdrawals, swaps.
  title: Trovo Wallet API
  version: "1.0"
paths:
...
  /v1/p2p/asset-classes:
    get:
      description: Returns every asset category the marketplace filter can narrow
        by (e.g. stablecoin, token). Small, mostly-static reference list.
      produces:
      - application/json
      responses:
        "200":
          description: 'data: list of asset classes'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - SignatureAuth: []
      summary: List P2P asset categories
      tags:
      - P2P
...
  /v1/p2p/marketplace/facets:
    get:
      description: Returns the distinct asset and currency values currently offered
        on the marketplace, for populating filter dropdowns.
      produces:
      - application/json
      responses:
        "200":
          description: Facet lists
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - SignatureAuth: []
      summary: List available P2P marketplace filter options
      tags:
      - P2P
...
  /v1/p2p/my-offers:
    get:
      description: Lists every offer the caller owns as merchant, in any state (active,
        paused, closed).
      produces:
      - application/json
      responses:
        "200":
          description: 'data: list of offers'
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - SignatureAuth: []
      summary: List my own P2P offers
      tags:
      - P2P
...
  /v1/p2p/offers:
    get:
      description: Lists active offers on the P2P marketplace, with optional filters
        and pagination. Used to populate the marketplace buy/sell listing screen.
      parameters:
      - description: Filter by BUY or SELL
        in: query
        name: offerType
        type: string
      - description: Filter by asset code
        in: query
        name: asset
        type: string
      - description: Filter by asset class ID
        in: query
        name: assetClassId
        type: integer
      - description: Filter by country code
        in: query
        name: countryCode
        type: string
      - description: Filter by fiat currency
        in: query
        name: currency
        type: string
      - description: Page number (default 1)
        in: query
        name: page
        type: integer
      - description: Results per page (default 20)
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: data (offers), total, page, pageSize
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      security:
      - SignatureAuth: []
      summary: Browse P2P marketplace offers
      tags:
      - P2P
    post:
      consumes:
      - application/json
      description: Creates a merchant offer to buy or sell an asset for fiat on the
        P2P marketplace. The caller becomes the merchant on the offer.
      parameters:
      - description: 'Offer details: offerType (BUY/SELL), asset, paymentMethodId
          (an existing saved payment method), merchantPayoutAddress, country, countryCode,
          currency, priceType, price, priceMargin, minOrderAmount, maxOrderAmount,
          availableLiquidity, remark'
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/p2p.createOfferRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created offer
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid JSON or validation error
          schema:
            additionalProperties: true
            type: object
      security:
      - SignatureAuth: []
      summary: Create a new P2P buy/sell offer
      tags:
      - P2P
  /v1/p2p/offers/{offerID}:
    get:
      parameters:
      - description: Offer ID
        in: path
        name: offerID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Offer not found
          schema:
            additionalProperties: true
            type: object
      security:
      - SignatureAuth: []
      summary: Get a single P2P offer by ID
      tags:
      - P2P
    put:
      consumes:
      - application/json
      description: Updates the editable fields of an offer the caller owns as merchant
        (price, limits, payout address, payment method, remark). Offer type and asset
        cannot be changed after creation.
      parameters:
      - description: Offer ID
        in: path
        name: offerID
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/p2p.updateOfferRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Updated offer
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid JSON or validation error
          schema:
            additionalProperties: true
            type: object
      security:
      - SignatureAuth: []
      summary: Update an existing P2P offer
      tags:
      - P2P
  /v1/p2p/offers/{offerID}/activate:
    post:
      description: Makes a paused offer visible and orderable on the marketplace again.
      parameters:
      - description: Offer ID
        in: path
        name: offerID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Not the offer's owner, or invalid state
          schema:
            additionalProperties: true
            type: object
      security:
      - SignatureAuth: []
      summary: Activate a paused P2P offer
      tags:
      - P2P
  /v1/p2p/offers/{offerID}/close:
    post:
      description: Closes an offer the caller owns as merchant. A closed offer stops
        appearing on the marketplace and cannot be reactivated (unlike pausing).
      parameters:
      - description: Offer ID
        in: path
        name: offerID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Closed offer
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Not the offer's owner, or invalid state
          schema:
            additionalProperties: true
            type: object
      security:
      - SignatureAuth: []
      summary: Permanently close a P2P offer
      tags:
      - P2P
  /v1/p2p/offers/{offerID}/pause:
    post:
      description: Temporarily hides an offer from the marketplace without closing
        it permanently. Can be reactivated later.
      parameters:
      - description: Offer ID
        in: path
        name: offerID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Not the offer's owner, or invalid state
          schema:
            additionalProperties: true
            type: object
      security:
      - SignatureAuth: []
      summary: Pause a P2P offer
      tags:
      - P2P
  /v1/p2p/offers/{offerID}/quote:
    get:
      description: Returns an itemized breakdown (fees, VAT, charges, net amounts
        for both buyer and seller) for a hypothetical order against this offer at
        the given amount, without creating or persisting anything. Used by the order-creation
        screen to show the customer real numbers before they commit.
      parameters:
      - description: Offer ID
        in: path
        name: offerID
        required: true
        type: string
      - description: Specified asset amount to quote
        in: query
        name: amount
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Fee breakdown
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      security:
      - SignatureAuth: []
      summary: Get a fee quote for an offer before creating an order
      tags:
      - P2P
...
securityDefinitions:
  BearerAuth:
    description: JWT bearer token used by the internal "trovo-manager" (tm-api admin)
      routes. Send as "Bearer <token>". See internal/middleware/authentication_middleware.go's
      JwtTokenAuthMiddleware.
    in: header
    name: Authorization
    type: apiKey
  ServiceLinkApiKey:
    description: Static per-partner API key used by white-label "service link" integrations
      (and by tm-api for a few admin/server-to-server calls). Issued when a service
      link is provisioned; see internal/components/servicelinks. See internal/middleware/api_key_middleware.go.
    in: header
    name: X-TW-SERVICE-LINK-API-KEY
    type: apiKey
  SignatureAuth:
    description: 'Request-signing scheme used by app-web and app-mobile for end-user
      requests. The caller sends four headers together: X-TW-PUBLIC-KEY (the wallet
      address making the call), X-TW-SIGNER (the address whose private key signed
      the request), X-TW-TIMESTAMP (a fresh timestamp), and X-TW-SIGNATURE (a signature
      over signer+timestamp, and for POST/PUT/PATCH over the request body/URI). There
      is no bearer token: every request is independently signed. See internal/middleware/authentication_middleware.go.'
    in: header
    name: X-TW-SIGNATURE
    type: apiKey
swagger: "2.0"